slug-classifier/scripts/train_binary.py
#!/usr/bin/env python3
"""
Training script for the binary classifier model.

This script should:
1. Parse command line arguments and configuration
2. Set up logging and experiment tracking
3. Load and preprocess data
4. Initialize the binary classifier model
5. Train the model with proper validation and checkpointing
6. Evaluate the model on test data
7. Save model artifacts and performance metrics

Example usage:
    python scripts/train_binary.py --config configs/binary_classifier.yaml
"""

import argparse
import logging
import os
import sys
import yaml
from pathlib import Path

# TODO: Import project modules

def parse_args():
    """Parse command line arguments."""
    parser = argparse.ArgumentParser(description='Train binary classifier')
    parser.add_argument('--config', type=str, required=True, help='Path to config file')
    parser.add_argument('--data_dir', type=str, help='Path to data directory (overrides config)')
    parser.add_argument('--output_dir', type=str, help='Path to output directory (overrides config)')
    parser.add_argument('--epochs', type=int, help='Number of epochs (overrides config)')
    parser.add_argument('--batch_size', type=int, help='Batch size (overrides config)')
    parser.add_argument('--seed', type=int, default=42, help='Random seed')
    parser.add_argument('--debug', action='store_true', help='Enable debug mode')
    
    return parser.parse_args()

def main():
    """Main training function."""
    # Parse arguments and load config
    args = parse_args()
    
    # TODO: Load configuration from YAML file
    
    # TODO: Set up logging and experiment tracking
    
    # TODO: Set random seeds for reproducibility
    
    # TODO: Prepare datasets and data loaders
    
    # TODO: Initialize model, optimizer, and scheduler
    
    # TODO: Training loop with validation
    #       - Track metrics
    #       - Implement early stopping
    #       - Save checkpoints
    
    # TODO: Evaluate on test set
    
    # TODO: Save final model and metrics

if __name__ == '__main__':
    main()
