slug-classifier/requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
pandas>=1.3.0
Pillow>=8.3.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=6.0
tqdm>=4.62.0

# Data processing
requests>=2.26.0
pyarrow>=5.0.0

# Model evaluation
torchmetrics>=0.5.0

# Visualization
plotly>=5.3.0

# API
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# Development
jupyter>=1.0.0
pytest>=6.2.5
black>=21.8b0
isort>=5.9.0
pylint>=2.10.0
